fix: delete

+++ src/main/java/com/devsuperior/dsCatagog/services/CategoryService.java
@@ -55,13 +55,13 @@ public class CategoryService {
     }
 
     public void delete(Long id) {
-       try{
-       repository.deleteById(id);
+    if(!repository.existsById(id)){
+        throw new ResourceNotFoundException("Recurso não encontrado");
     }
-       catch (EmptyResultDataAccessException e){
-           throw new ResourceNotFoundException("Id not found"+ id);
-       }catch (DataIntegrityViolationException e ){
-         throw new DatabaseException("Violação de integridade");
-       }
+    try{
+        repository.deleteById(id);
+    }catch (DataIntegrityViolationException e ){
+        throw new DatabaseException("Falha de integridade referencial");
+    }
 }}
